[Fix] #582 - 기존 파일과 분리
[Fix] #582 - 기존 파일과 분리
