+++ WSSiOS/WSSiOS/Source/Data/DTO/User/UserNovelListResponse.swift
@@ -9,6 +9,7 @@ import Foundation
 
 struct UserNovelListResponse: Decodable {
     let userNovelCount: Int
+    let userNovelRating: Float
     let isLoadable: Bool
     let userNovels: [UserNovelResponse]
 }
@@ -16,15 +17,8 @@ struct UserNovelListResponse: Decodable {
 struct UserNovelResponse: Decodable {
     let userNovelId: Int
     let novelId: Int
-    let title: String
+    let author: String
     let novelImage: String
+    let title: String
     let novelRating: Float
-    let readStatus: String?
-    let isInterest: Bool
-    let userNovelRating: Float
-    let attractivePoints: [String]
-    let startDate: String?
-    let endDate: String?
-    let keywords: [String]
-    let myFeeds: [String]
 }
